Create build script to create constants needed for RecursiveGaussian
